hacky spawning fix

+++ plugin_manager/plugin_manager.go
@@ -6,6 +6,7 @@ import (
 	"os"
 	"os/exec"
 	"sync"
+	"syscall"
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/turbot/steampipe-plugin-sdk/logging"
@@ -62,11 +63,14 @@ func (m *PluginManager) Get(req *pb.GetRequest) (resp *pb.GetResponse, err error
 		}
 	}()
 
-	log.Printf("[WARN] ****************** PluginManager %p Get connection '%s'\n", m, req.Connection)
+	log.Printf("[WARN] ****************** PluginManager %p Get connection '%s', plugins %+v\n", m, req.Connection, m.Plugins)
+	return &pb.GetResponse{
+		Reattach: &pb.ReattachConfig{Protocol: "TEST"},
+	}, nil
 
 	// is this plugin already running
 	if plugin, ok := m.Plugins[req.Connection]; ok {
-		log.Printf("[TRACE] found '%s' in map %v", req.Connection, m.Plugins)
+		log.Printf("[WARN] PluginManager %p found '%s' in map %v", m, req.Connection, m.Plugins)
 
 		// check the pid exists
 		exists, _ := utils.PidExists(int(plugin.Pid))
@@ -76,12 +80,12 @@ func (m *PluginManager) Get(req *pb.GetRequest) (resp *pb.GetResponse, err error
 				Reattach: plugin,
 			}, nil
 		}
-		log.Printf("[WARN] plugin pid %d for connection '%s' found in plugin map but does not exist - removing from map", plugin.Pid, req.Connection)
+		//log.Printf("[WARN] PluginManager %p plugin pid %d for connection '%s' found in plugin map but does not exist - removing from map", m, plugin.Pid, req.Connection)
 		// so there is an entry in the map but it does not exist - remove from the map
 		delete(m.Plugins, req.Connection)
 	}
 
-	log.Printf("[TRACE] '%s' NOT found in map %v - starting", req.Connection, m.Plugins)
+	log.Printf("[WARN] PluginManager %p '%s' NOT found in map %v - starting", m, req.Connection, m.Plugins)
 	// so we need to start the plugin
 	reattach, err := m.startPlugin(req)
 	if err != nil {
@@ -90,6 +94,8 @@ func (m *PluginManager) Get(req *pb.GetRequest) (resp *pb.GetResponse, err error
 
 	// store the reattach config in our map
 	m.Plugins[req.Connection] = reattach
+
+	log.Printf("[WARN] ****************** PluginManager %p Get complete", m)
 
 	// and return
 	return &pb.GetResponse{
@@ -110,7 +116,7 @@ func (m *PluginManager) SetConnectionConfigMap(req *pb.SetConnectionConfigMapReq
 	return &pb.SetConnectionConfigMapResponse{}, nil
 }
 
-func (m *PluginManager) Shutdown(req *pb.ShutdownRequest) (resp *pb.ShutdownResponse, err error) {
+func (m *PluginManager) Shutdown(*pb.ShutdownRequest) (resp *pb.ShutdownResponse, err error) {
 	m.mut.Lock()
 	defer func() {
 		m.mut.Unlock()
@@ -119,8 +125,12 @@ func (m *PluginManager) Shutdown(req *pb.ShutdownRequest) (resp *pb.ShutdownResp
 		}
 	}()
 
+	log.Printf("[WARN] Shutdown")
+	return &pb.ShutdownResponse{}, nil
+
 	var errs []error
 	for _, p := range m.Plugins {
+		log.Printf("[WARN] kill %v", p)
 		err = m.killPlugin(p.Pid)
 		if err != nil {
 			errs = append(errs, err)
@@ -155,6 +165,11 @@ func (m *PluginManager) startPlugin(req *pb.GetRequest) (*pb.ReattachConfig, err
 	cmd := exec.Command(pluginPath)
 	// pass env to command
 	cmd.Env = os.Environ()
+	// set attributes on the command to ensure the process is not shutdown when its parent terminates
+	cmd.SysProcAttr = &syscall.SysProcAttr{
+		//Setpgid:    true,
+		Foreground: false,
+	}
 	client := plugin.NewClient(&plugin.ClientConfig{
 		HandshakeConfig:  sdkpluginshared.Handshake,
 		Plugins:          pluginMap,
@@ -168,12 +183,39 @@ func (m *PluginManager) startPlugin(req *pb.GetRequest) (*pb.ReattachConfig, err
 	}
 	reattach := client.ReattachConfig()
 	return pb.NewReattachConfig(reattach), nil
+
+	/* hub did this
+
+	// loop as we may need to retry if the plugin exists in the map but has actually exited
+	const maxAttempts = 3
+	for attempt := 1; attempt < maxAttempts; attempt++ {
+		// ask connection map to get or create this connection
+		c, err := h.connections.get(pluginFQN, connectionName)
+		if err != nil {
+			return nil, err
+		}
+
+		// make sure that the plugin is running
+		// (i.e. it has not crashed)
+		if !c.Plugin.Client.Exited() {
+			// it is running, return it
+			return c, nil
+		}
+
+		// remove connection from the connection map and kill the GRPC client
+		h.connections.removeAndKill(pluginFQN, connectionName)
+	}
+	*/
 }
 
 func (m *PluginManager) killPlugin(pid int64) error {
 	process, err := utils.FindProcess(int(pid))
 	if err != nil {
+		log.Printf("[WARN] error finding process %d", pid)
 		return err
+	}
+	if process == nil {
+		return nil
 	}
 	return process.Kill()
 }
